refactor(chat): extract @endo/monaco-wrapper and @endo/space-file-explorer
Move the Monaco editor wrapper and the File Explorer Space out of the
chat app into two standalone workspace packages:

- @endo/monaco-wrapper — the lazy-loaded monaco-editor wrapper
  (detectTheme, createMonacoEditor, colorize), previously chat's
  monaco-wrapper.js. Its 13 chat import sites are rewired to the package
  specifier.
- @endo/space-file-explorer — file-explorer{,-fs,-component,-env}.js
  plus its private layer-diff.js. chat now imports fileExplorerComponent
  from the package, which depends on @endo/monaco-wrapper so the former
  chat <-> file-explorer monaco coupling no longer forms a cycle.

Both packages follow the existing Space-package convention (whylip /
preact-container): a root index.js barrel over src/, a browser
eslintConfig, and exclusion from the root tsconfig like the other
browser/UI packages. Styling still lives in the chat stylesheet; a
self-contained CSS export is a follow-up.
