+++ packages/chat/package.json
@@ -58,11 +58,13 @@
     "@endo/init": "workspace:^",
     "@endo/inventory-graph": "workspace:^",
     "@endo/markmdown": "workspace:^",
+    "@endo/monaco-wrapper": "workspace:^",
     "@endo/pass-style": "workspace:^",
     "@endo/patterns": "workspace:^",
     "@endo/platform": "workspace:^",
     "@endo/preact-container": "workspace:^",
     "@endo/promise-kit": "workspace:^",
+    "@endo/space-file-explorer": "workspace:^",
     "@endo/where": "workspace:^",
     "@endo/whylip": "workspace:^",
     "marked": "^17.0.5",
