+++ packages/chat/test/component/blob-viewer.test.js
@@ -41,7 +41,7 @@ register(
 );
 
 // Dynamically imported AFTER the loader is registered so blob-viewer's (and
-// markdown-preview's) `import ... from './monaco-wrapper.js'` resolve to the
+// markdown-preview's) `import ... from '@endo/monaco-wrapper'` resolve to the
 // stub.
 const { createBlobViewer } = await import('../../blob-viewer.js');
