+++ packages/chat/test/component/counter-proposal-form.test.js
@@ -45,7 +45,7 @@ register(
 );
 
 // Dynamically imported AFTER the loader is registered so the form's
-// `import { createMonacoEditor } from './monaco-wrapper.js'` resolves to the
+// `import { createMonacoEditor } from '@endo/monaco-wrapper'` resolves to the
 // stub.
 const { createCounterProposalForm } =
   await import('../../counter-proposal-form.js');
