+++ packages/chat/test/component/define-form.test.js
@@ -41,7 +41,7 @@ register(
 );
 
 // Dynamically imported AFTER the loader is registered so define-form's
-// `import { createMonacoEditor } from './monaco-wrapper.js'` resolves to the
+// `import { createMonacoEditor } from '@endo/monaco-wrapper'` resolves to the
 // stub.
 const { createDefineForm } = await import('../../define-form.js');
