+++ packages/monaco-wrapper/index.js
@@ -1,0 +1,7 @@
+// @ts-check
+
+export {
+  detectTheme,
+  createMonacoEditor,
+  colorize,
+} from './src/monaco-wrapper.js';
