+++ packages/monaco-wrapper/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@endo/monaco-wrapper",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Lazy-loaded Monaco editor wrapper: theme detection, confined editor creation, and standalone colorize",
+  "keywords": [
+    "endo",
+    "monaco",
+    "editor"
+  ],
+  "author": "Endo contributors",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/endojs/endo.git",
+    "directory": "packages/monaco-wrapper"
+  },
+  "bugs": {
+    "url": "https://github.com/endojs/endo/issues"
+  },
+  "type": "module",
+  "exports": {
+    ".": "./index.js",
+    "./package.json": "./package.json"
+  },
+  "scripts": {
+    "build": "exit 0",
+    "lint": "yarn lint:eslint",
+    "lint-fix": "yarn lint:eslint --fix",
+    "lint:eslint": "eslint '**/*.js'"
+  },
+  "dependencies": {
+    "@endo/harden": "workspace:^",
+    "monaco-editor": "^0.55.1"
+  },
+  "devDependencies": {
+    "eslint": "catalog:dev"
+  },
+  "eslintConfig": {
+    "extends": [
+      "plugin:@endo/internal"
+    ],
+    "env": {
+      "browser": true
+    }
+  }
+}
