+++ packages/space-file-explorer/index.js
@@ -1,0 +1,3 @@
+// @ts-check
+
+export { fileExplorerComponent } from './src/file-explorer-component.js';
