+++ packages/space-file-explorer/package.json
@@ -1,0 +1,51 @@
+{
+  "name": "@endo/space-file-explorer",
+  "version": "0.1.0",
+  "private": true,
+  "description": "The File Explorer Space: browse and edit Endo filesystem, mount, and exo-git capabilities in the chat client",
+  "keywords": [
+    "endo",
+    "file-explorer",
+    "filesystem",
+    "space"
+  ],
+  "author": "Endo contributors",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/endojs/endo.git",
+    "directory": "packages/space-file-explorer"
+  },
+  "bugs": {
+    "url": "https://github.com/endojs/endo/issues"
+  },
+  "type": "module",
+  "exports": {
+    ".": "./index.js",
+    "./package.json": "./package.json"
+  },
+  "scripts": {
+    "build": "exit 0",
+    "lint": "yarn lint:eslint",
+    "lint-fix": "yarn lint:eslint --fix",
+    "lint:eslint": "eslint '**/*.js'"
+  },
+  "dependencies": {
+    "@endo/exo-stream": "workspace:^",
+    "@endo/far": "workspace:^",
+    "@endo/harden": "workspace:^",
+    "@endo/monaco-wrapper": "workspace:^",
+    "@endo/platform": "workspace:^"
+  },
+  "devDependencies": {
+    "eslint": "catalog:dev"
+  },
+  "eslintConfig": {
+    "extends": [
+      "plugin:@endo/internal"
+    ],
+    "env": {
+      "browser": true
+    }
+  }
+}
