+++ packages/space-file-explorer/src/file-explorer-component.js
