+++ packages/space-file-explorer/src/file-explorer-env.js
