+++ packages/space-file-explorer/src/file-explorer-fs.js
