+++ packages/space-file-explorer/src/layer-diff.js
