+++ tsconfig.json
@@ -29,8 +29,10 @@
         "packages/ses-test",
         "packages/jaine",
         "packages/lal",
+        "packages/monaco-wrapper",
         "packages/platform",
         "packages/preact-container",
+        "packages/space-file-explorer",
         "packages/whylip",
         "**/build",
         "**/demo",
